+++ src/audiovideocontroller/DefaultAudioVideoController.ts
@@ -3,6 +3,7 @@
 
 import ActiveSpeakerDetector from '../activespeakerdetector/ActiveSpeakerDetector';
 import DefaultActiveSpeakerDetector from '../activespeakerdetector/DefaultActiveSpeakerDetector';
+import Attendee from '../attendee/Attendee';
 import AudioMixController from '../audiomixcontroller/AudioMixController';
 import DefaultAudioMixController from '../audiomixcontroller/DefaultAudioMixController';
 import AudioVideoController from '../audiovideocontroller/AudioVideoController';
@@ -29,6 +30,7 @@ import SessionStateControllerState from '../sessionstatecontroller/SessionStateC
 import SessionStateControllerTransitionResult from '../sessionstatecontroller/SessionStateControllerTransitionResult';
 import DefaultSignalingClient from '../signalingclient/DefaultSignalingClient';
 import { SdkStreamServiceType } from '../signalingprotocol/SignalingProtocol.js';
+import SimulcastLayers from '../simulcastlayers/SimulcastLayers';
 import DefaultStatsCollector from '../statscollector/DefaultStatsCollector';
 import AttachMediaInputTask from '../task/AttachMediaInputTask';
 import CleanRestartedSessionTask from '../task/CleanRestartedSessionTask';
@@ -64,13 +66,15 @@ import SimulcastVideoStreamIndex from '../videostreamindex/SimulcastVideoStreamI
 import DefaultVideoTileController from '../videotilecontroller/DefaultVideoTileController';
 import VideoTileController from '../videotilecontroller/VideoTileController';
 import DefaultVideoTileFactory from '../videotilefactory/DefaultVideoTileFactory';
+import DefaultSimulcastUplinkPolicy from '../videouplinkbandwidthpolicy/DefaultSimulcastUplinkPolicy';
 import NScaleVideoUplinkBandwidthPolicy from '../videouplinkbandwidthpolicy/NScaleVideoUplinkBandwidthPolicy';
-import SimulcastUplinkPolicy from '../videouplinkbandwidthpolicy/SimulcastUplinkPolicy';
+import SimulcastUplinkObserver from '../videouplinkbandwidthpolicy/SimulcastUplinkObserver';
 import DefaultVolumeIndicatorAdapter from '../volumeindicatoradapter/DefaultVolumeIndicatorAdapter';
 import WebSocketAdapter from '../websocketadapter/WebSocketAdapter';
 import AudioVideoControllerState from './AudioVideoControllerState';
 
-export default class DefaultAudioVideoController implements AudioVideoController {
+export default class DefaultAudioVideoController
+  implements AudioVideoController, SimulcastUplinkObserver {
   private _logger: Logger;
   private _configuration: MeetingSessionConfiguration;
   private _webSocketAdapter: WebSocketAdapter;
@@ -237,10 +241,12 @@ export default class DefaultAudioVideoController implements AudioVideoController
 
     this.meetingSessionContext.enableSimulcast = this.enableSimulcast;
     if (this.enableSimulcast) {
-      this.meetingSessionContext.videoUplinkBandwidthPolicy = new SimulcastUplinkPolicy(
+      const simulcastPolicy = new DefaultSimulcastUplinkPolicy(
         this.configuration.credentials.attendeeId,
         this.meetingSessionContext.logger
       );
+      simulcastPolicy.addObserver(this);
+      this.meetingSessionContext.videoUplinkBandwidthPolicy = simulcastPolicy;
       this.meetingSessionContext.videoDownlinkBandwidthPolicy = new VideoAdaptiveProbePolicy(
         this.logger,
         this.meetingSessionContext.videoTileController
@@ -738,5 +744,23 @@ export default class DefaultAudioVideoController implements AudioVideoController
     if (!!this.meetingSessionContext && this.meetingSessionContext.signalingClient) {
       this.meetingSessionContext.signalingClient.resume([streamId]);
     }
+  }
+
+  getRemoteVideosAvailable(): Attendee[] {
+    const { videoStreamIndex } = this.meetingSessionContext;
+    if (!videoStreamIndex) {
+      this.logger.info('meeting has not started');
+      return [];
+    }
+    const selfAttendeeId = this.configuration.credentials.attendeeId;
+    return videoStreamIndex.allVideoSendingAttendeesExcludingSelf(selfAttendeeId);
+  }
+
+  encodingSimulcastLayerDidChange(simulcastLayers: SimulcastLayers): void {
+    this.forEachObserver(observer => {
+      Maybe.of(observer.encodingSimulcastLayerDidChange).map(f =>
+        f.bind(observer)(simulcastLayers)
+      );
+    });
   }
 }