+++ src/audiovideoobserver/AudioVideoObserver.ts
@@ -1,11 +1,13 @@
 // Copyright 2019-2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+import Attendee from '../attendee/Attendee';
 import ClientMetricReport from '../clientmetricreport/ClientMetricReport';
 import ClientVideoStreamReceivingReport from '../clientmetricreport/ClientVideoStreamReceivingReport';
 import ConnectionHealthData from '../connectionhealthpolicy/ConnectionHealthData';
 import MeetingSessionStatus from '../meetingsession/MeetingSessionStatus';
 import MeetingSessionVideoAvailability from '../meetingsession/MeetingSessionVideoAvailability';
+import SimulcastLayers from '../simulcastlayers/SimulcastLayers';
 import VideoTileState from '../videotile/VideoTileState';
 
 export default interface AudioVideoObserver {
@@ -70,7 +72,7 @@ export default interface AudioVideoObserver {
   ): void;
 
   /**
-   * Called when one or more remote video streams do not meet expected average bitrate
+   * Called when one or more remote video streams do not meet expected average bitrate.
    */
   videoNotReceivingEnoughData?(receivingDataMap: ClientVideoStreamReceivingReport[]): void;
 
@@ -107,4 +109,14 @@ export default interface AudioVideoObserver {
    * trigger a message to the user about the situation.
    */
   videoSendDidBecomeUnavailable?(): void;
+
+  /**
+   * Called when the remote attendees sending video gets changed.
+   */
+  remoteVideosAvailableDidChange?(attendees: Attendee[]): void;
+
+  /**
+   * Called when simulcast uplink encoding renditions are changed.
+   */
+  encodingSimulcastLayerDidChange?(simulcastLayers: SimulcastLayers): void;
 }